Duke3d: don't show button double press options for controllers in the menu
The support will remain internally, but these aren't useful for the vast majority of users using modern controllers, serving only to overwhelm them with options that are of limited use. This has the bonus of letting the entirety of the standard Xbox controller's buttons fit on one screen without scrolling.